requirements.txt
asgiref==3.8.1
babel==2.16.0
crispy-bootstrap5==2024.2
Django==5.1
django-bootstrap5==24.2
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-js-asset==2.2.0
django-money==3.5.3
django-phone-field==1.8.1
django-taggit==6.0.0
phonenumbers==8.13.46
pillow==10.4.0
py-moneyed==3.0
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
